TST: fix assert_categorical_equal message
stage 1 of #13076

Author: sinhrks <sinhrks@gmail.com>

Closes #13080 from sinhrks/test_categorical_message and squashes the following commits:

81172ce [sinhrks] TST: fix assert_categorical_equal message

+++ pandas/core/categorical.py
@@ -985,7 +985,7 @@ class Categorical(PandasObject):
 
         # Provide compatibility with pre-0.15.0 Categoricals.
         if '_codes' not in state and 'labels' in state:
-            state['_codes'] = state.pop('labels')
+            state['_codes'] = state.pop('labels').astype(np.int8)
         if '_categories' not in state and '_levels' in state:
             state['_categories'] = self._validate_categories(state.pop(
                 '_levels'))
